:sparkles:[feat]: useModal hook생성
