+++ src/app/hooks/upload/useModal.ts
@@ -1,0 +1,27 @@
+'use client';
+import { useState } from 'react';
+
+export const useModal = (initialState = false) => {
+  const [isOpen, setIsOpen] = useState(initialState);
+
+  const openModal = () => {
+    setIsOpen(true);
+  };
+
+  const closeModal = () => {
+    setIsOpen(false);
+  };
+
+  const handleModalClick = (
+    event: React.MouseEvent<HTMLDivElement, MouseEvent>,
+  ) => {
+    event.stopPropagation();
+  };
+
+  return {
+    isOpen,
+    openModal,
+    closeModal,
+    handleModalClick,
+  };
+};
